refactored all codebase (unfinished)
changing devices, so i am committing it to pull it on another device
